refactor(wallet-sdk): change type imports
