+++ sdk/wallet-sdk/src/wallet/init/index.ts
@@ -1,6 +1,7 @@
 // Copyright (c) 2025-2026 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+export * from './types/context.js'
 export * from './initializedSDK.js'
 export * from './types/index.js'
 export { SDKPlugin } from './plugin.js'
