+++ sdk/wallet-sdk/src/wallet/namespace/events/types.ts
@@ -4,7 +4,7 @@
 import { AuthTokenProvider } from '@canton-network/core-wallet-auth'
 import { PartyId } from '@canton-network/core-types'
 import { type LedgerCommonSchemas } from '@canton-network/core-ledger-client-types'
-import { SDKContext } from '../../sdk.js'
+import type { SDKContext } from '../../init/types/context.js'
 import { ParsedURL } from '../utils/url.js'
 
 export type UpdatesOptions = {
