+++ sdk/wallet-sdk/src/wallet/namespace/transactions/prepared.ts
@@ -6,7 +6,7 @@ import {
     signTransactionHash,
 } from '@canton-network/core-signing-lib'
 import { SignedTransaction } from './signed.js'
-import { SDKContext } from '../../sdk.js'
+import type { SDKContext } from '../../init/types/context.js'
 import { Ops } from '@canton-network/core-provider-ledger'
 import { decodePreparedTransaction } from '@canton-network/core-tx-visualizer'
 import { LedgerNamespace } from '../ledger/index.js'
