+++ sdk/wallet-sdk/src/wallet/namespace/transactions/types.ts
@@ -1,7 +1,7 @@
 // Copyright (c) 2025-2026 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { LedgerTypes } from '../../sdk.js'
+import type { LedgerCommonSchemas } from '@canton-network/core-ledger-client-types'
 import { RawCommandMap, WrappedCommand } from '../ledger/index.js'
 
 export type PreparedCommand<
@@ -15,5 +15,5 @@ export type PreparedCommand<
         : K extends keyof RawCommandMap
           ? WrappedCommand<K>
           : never,
-    LedgerTypes['DisclosedContract'][],
+    LedgerCommonSchemas['DisclosedContract'][],
 ]
