+++ sdk/wallet-sdk/src/wallet/namespace/utils/url.ts
@@ -1,7 +1,7 @@
 // Copyright (c) 2025-2026 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { SDKContext } from '../../sdk.js'
+import type { SDKContext } from '../../init/types/context.js'
 import { TokenStandardService } from '@canton-network/core-token-standard-service'
 
 export type URLInput = URL | string
